+++ src/plugins_types/lyds_tree.c
@@ -27,7 +27,7 @@ static void lyplg_type_free_lyds(const struct ly_ctx *ctx, struct lyd_value *val
 
 static LY_ERR
 lyplg_type_store_lyds(const struct ly_ctx *ctx, const struct lysc_type *type, const void *value,
-        size_t UNUSED(value_len), uint32_t options, LY_VALUE_FORMAT format, void *UNUSED(prefix_data),
+        uint32_t UNUSED(value_len), uint32_t options, LY_VALUE_FORMAT format, void *UNUSED(prefix_data),
         uint32_t UNUSED(hints), const struct lysc_node *UNUSED(ctx_node), struct lyd_value *storage,
         struct lys_glob_unres *UNUSED(unres), struct ly_err_item **UNUSED(err))
 {
@@ -107,7 +107,7 @@ lyplg_type_sort_lyds(const struct ly_ctx *UNUSED(ctx), const struct lyd_value *U
 
 static const void *
 lyplg_type_print_lyds(const struct ly_ctx *UNUSED(ctx), const struct lyd_value *UNUSED(value),
-        LY_VALUE_FORMAT UNUSED(format), void *UNUSED(prefix_data), ly_bool *dynamic, size_t *value_len)
+        LY_VALUE_FORMAT UNUSED(format), void *UNUSED(prefix_data), ly_bool *dynamic, uint32_t *value_len)
 {
     if (dynamic) {
         *dynamic = 0;
